+++ core/src/main/scala/mongo4s/internal/DirectAggregateQueryImpl.scala
@@ -5,6 +5,7 @@ import scala.concurrent.duration.FiniteDuration
 import org.bson.conversions.Bson
 import org.bson.{BsonDocument, BsonInt32}
 import org.reactivestreams.Publisher
+import com.mongodb.ExplainVerbosity
 import com.mongodb.client.model.Collation
 import com.mongodb.reactivestreams.client.{AggregatePublisher, ClientSession, MongoCollection as RSMongoCollection}
 
@@ -31,6 +32,9 @@ private[mongo4s] final class DirectAggregateQueryImpl[F[*], S[*], A](
   def maxTime(duration: FiniteDuration): AggregateQuery[F, S, A] = copy(options = options.withMaxTime(duration))
   def batchSize(n: Int): AggregateQuery[F, S, A]                 = copy(options = options.withBatchSize(n))
   def comment(value: String): AggregateQuery[F, S, A]            = copy(options = options.withComment(value))
+
+  def explain(verbosity: ExplainVerbosity): F[BsonDocument] =
+    rs.one(documents(limited = false).explain(classOf[BsonDocument], verbosity))
 
   def first: F[Option[A]]                  = rs.option(typed(limited = true))
   def all: F[List[A]]                      = rs.list(typed(limited = false))