+++ core/src/main/scala/mongo4s/internal/DirectFindQueryImpl.scala
@@ -4,6 +4,7 @@ import scala.concurrent.duration.FiniteDuration
 
 import org.bson.BsonDocument
 import org.reactivestreams.Publisher
+import com.mongodb.ExplainVerbosity
 import com.mongodb.client.model.Collation
 import com.mongodb.reactivestreams.client.{ClientSession, FindPublisher, MongoCollection as RSMongoCollection}
 
@@ -38,6 +39,9 @@ private[mongo4s] final class DirectFindQueryImpl[F[*], S[*], A](
   def maxTime(duration: FiniteDuration): FindQuery[F, S, A] = copy(options = options.withMaxTime(duration))
   def batchSize(n: Int): FindQuery[F, S, A]                 = copy(options = options.withBatchSize(n))
   def comment(value: String): FindQuery[F, S, A]            = copy(options = options.withComment(value))
+
+  def explain(verbosity: ExplainVerbosity): F[BsonDocument] =
+    rs.one(configure(rawDocuments, limit).explain(classOf[BsonDocument], verbosity))
 
   def first: F[Option[A]]                  = rs.option(publisher(Some(1)))
   def all: F[List[A]]                      = rs.list(publisher(limit))