+++ core/src/main/scala/mongo4s/queries/AggregateQuery.scala
@@ -3,6 +3,7 @@ package mongo4s.queries
 import scala.concurrent.duration.FiniteDuration
 
 import org.bson.BsonDocument
+import com.mongodb.ExplainVerbosity
 
 import mongo4s.Streamable
 
@@ -14,6 +15,8 @@ trait AggregateQuery[F[*], S[*], A]:
   def maxTime(duration: FiniteDuration): AggregateQuery[F, S, A]
   def batchSize(n: Int): AggregateQuery[F, S, A]
   def comment(value: String): AggregateQuery[F, S, A]
+
+  def explain(verbosity: ExplainVerbosity = ExplainVerbosity.QUERY_PLANNER): F[BsonDocument]
 
   def first: F[Option[A]]
   def all: F[List[A]]